\name{gse64495}
\docType{data}
\alias{gse64495}
\title{A DNA methylation array with no missing values}
\description{
  This dataset is a subset of the Illumina HumanMethylation450 BeadChip Platform dataset with GEO label GSE64495.
  It contains beta values of 100 samples for a subset 200 CpGs without missing values.
}
\usage{gse64495}
\format{A DNA methylation array containing 200 CpG and 100 samples.}
\source{}
\references{Di Lena P, Sala C, Prodi A, Nardini C. Missing value estimation methods for DNA methylation data. submitted to Bioinformatics.
}
\keyword{datasets}